refactor: eliminate onAddCell prop drilling with useAddCell hook
- Create useAddCell hook handling cell creation with fractional indices
- Use schema's createCellBetween for expert fractional index management
- Remove onAddCell from all component interfaces (CellList, Cell, CellBetweener, etc)
- Simplify interfaces: CellList down to 1 prop, Cell down to 2 props
- Maintain AI model defaults, focus management, and all existing functionality
- Remove unused addCell callback and imports from NotebookViewer
- All 106 tests continue to pass
